+++ Source/tusdotnet.benchmark/Program.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Threading.Tasks;
-using BenchmarkDotNet.Running;
-
-//using tusdotnet.benchmark.Benchmarks;
+﻿//using tusdotnet.benchmark.Benchmarks;
 
 namespace tusdotnet.benchmark
 {
